Yasin

+++ app/Http/Controllers/Admin/AdminAmenityController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Amenity;
+use App\Models\PackageAmenity;
 
 class AdminAmenityController extends Controller
 {
@@ -56,6 +57,11 @@ class AdminAmenityController extends Controller
 
     public function delete($id)
     {
+        $total = PackageAmenity::where('amenity_id',$id)->count();
+        if($total>0) {
+            return redirect()->back()->with('error','Amenity id Assigned to Package(s), So it can not be deleted');
+        }
+
         $obj = Amenity::where('id', $id)->first();
         $obj->delete();
         return redirect()->route('admin_amenity_index')->with('success', 'Amenity is deleted successfully');
